refactor(waiting): 예약 승격 로직 WaitingCommandService의 책임으로 리팩터링
