+++ src/test/java/roomescape/service/ReservationCommandServiceTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.verify;
 
 import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
@@ -17,7 +16,6 @@ import org.springframework.test.context.jdbc.Sql;
 import roomescape.controller.FixedClockConfig;
 import roomescape.domain.Member;
 import roomescape.domain.Reservation;
-import roomescape.domain.Slot;
 import roomescape.exception.DuplicateException;
 import roomescape.exception.ResourceNotFoundException;
 import roomescape.repository.ReservationDao;
@@ -35,7 +33,7 @@ class ReservationCommandServiceTest {
 
     @Autowired
     private ReservationCommandService reservationCommandService;
-    @MockitoSpyBean
+    @Autowired
     private WaitingDao waitingDao;
     @Autowired
     private WaitingQueryService waitingQueryService;
@@ -91,7 +89,6 @@ class ReservationCommandServiceTest {
 
         assertThat(reservationDao.findAllByName(new Member("user_e"))).hasSize(1);
         assertThat(waitingDao.findById(3L)).isEmpty();
-        verify(waitingDao).findNextInLineForUpdate(any(Slot.class));
     }
 
     @Test
@@ -123,7 +120,6 @@ class ReservationCommandServiceTest {
 
         assertThat(reservationDao.findAllByName(new Member("user_e"))).hasSize(1);
         assertThat(waitingDao.findById(3L)).isEmpty();
-        verify(waitingDao).findNextInLineForUpdate(any(Slot.class));
     }
 
     @Test
@@ -157,7 +153,6 @@ class ReservationCommandServiceTest {
 
         assertThat(reservationDao.findAllByName(new Member("user_d"))).hasSize(1);
         assertThat(waitingDao.findById(2L)).isEmpty();
-        verify(waitingDao).findNextInLineForUpdate(any(Slot.class));
     }
 
     @Test
